refactor(cli): extract versions commands to bin/commands/versions.js
Move `versions`, `version-delete`, and `versions-purge` into a dedicated
module. No behavior changes — same prompts, exit codes, and analytics keys.
